feat(context): add the fleet information in Ec2Selection
Add the Ec2Selection.Fleet field to store what fleet is selected. If the
spec is an instance spec, then store the fleets of all selected
instances (without duplicate). If the spec is about fleet, then store
every matching fleets, whether they have instances or not.
In both cases, there is no particular order for fleets.