fix(launch): fix aws launch time format
AWS changed its way of handling the ValidUntil parameter for the launch
requests. Somehow it now requires that the time.Time object comes from
a RFC3339 parsing.
Ec2tools now formats the time parameter of a launch request in RFC3339
then parses it again before to give it to AWS.